Changed expiration to time *or* views, whatever one comes first.
Defined maximum time and views for passwords: 90 days, 90 views.

Showing the amount of views left and the date of expiration on the view screen.

Changed site_url/password to site_url/gen and site_url/securesend to site_url/send (routes for each of them)

Removed references to MFYU from the code, changed to GNSND.